signed/unsigned warning fix. Added some double related shader functions
git-svn-id: https://www.h3d.org:8090/H3DAPI/trunk/H3DAPI@2375 610dfd11-2700-0410-9b68-9ae28eac5983

+++ src/ShaderPart.cpp
@@ -187,7 +187,7 @@ void ShaderPart::SFShaderString::update() {
         is.seekg (0, ios::beg);
         
       // allocate memory:
-        buffer = new char [length + 1];
+        buffer = new char [(unsigned int)length + 1];
         // read data as a block:
         is.read (buffer,length);
         length = is.gcount();
